+++ dandy/cache/decorators.py
@@ -1,8 +1,8 @@
 from dandy.cache.cache import BaseCache
 from dandy.cache.events import CacheEvent
 from dandy.cache.utils import generate_hash_key
-from dandy.debug import DebugRecorder
-from dandy.debug.utils import generate_new_debug_event_id
+from dandy.recorder import Recorder
+from dandy.recorder.utils import generate_new_recorder_event_id
 from dandy.intel import BaseIntel
 
 
@@ -21,15 +21,15 @@ def cache_decorator_function(
     cached_value = cache.get(hash_key)
 
     if cached_value:
-        if DebugRecorder.is_recording:
+        if Recorder.is_recording:
             if isinstance(cached_value, BaseIntel):
                 response = cached_value.model_dump_json(indent=4)
             else:
                 response = str(cached_value)
 
-            DebugRecorder.add_event(CacheEvent(
+            Recorder.add_event(CacheEvent(
                 response=response,
-                id=generate_new_debug_event_id()
+                id=generate_new_recorder_event_id()
             ))
 
         return cached_value