+++ dandy/recorder/renderer/json.py
@@ -1,12 +1,6 @@
-from pydantic import BaseModel
-
-from dandy.constants import __VERSION__
-
-from datetime import datetime
 from pathlib import Path
 
 from dandy.recorder.renderer.renderer import BaseRecordingRenderer
-from dandy.recorder.utils import generate_new_recorder_event_id
 
 
 class JsonRecordingRenderer(BaseRecordingRenderer):
@@ -15,12 +9,11 @@ class JsonRecordingRenderer(BaseRecordingRenderer):
 
     def to_file(
             self,
-            path: str,
-            file_name: str,
+            path: Path
     ):
         Path(path).mkdir(parents=True, exist_ok=True)
 
-        with open(Path(path, file_name), 'w') as new_file:
+        with open(Path(path, f'{self.recording.name}_recording_output.json'), 'w') as new_file:
             new_file.write(self.to_str())
 
     def to_str(self) -> str: